# backend/api/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação de API da aplicação Intellecta",
        "version": "0.1"
    },
    "paths": {
...
        "/institutions/{institution_id}/classes/{class_id}/users": {
            "get": {
                "tags": [
                    "Class Users"
                ],
                "summary": "Get all users in a class",
                "operationId": "getClassUsers",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of class users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClassUserResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Institution or class not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Class Users"
                ],
                "summary": "Add users to a class",
                "operationId": "createClassUsers",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User IDs to add to the class",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_ids"
                                ],
                                "properties": {
                                    "user_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "404": {
                        "description": "Institution or class not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/institutions/{institution_id}/classes/{class_id}/users/{class_users_id}": {
            "get": {
                "tags": [
                    "Class Users"
                ],
                "summary": "Get a specific user in a class",
                "operationId": "getClassUserById",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_users_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Class user details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassUserResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found in class"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Class Users"
                ],
                "summary": "Remove a user from a class",
                "operationId": "removeClassUser",
                "parameters": [
                    {
                        "name": "institution_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "class_users_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User removed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found in class"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClassUserResponse": {
                "properties": {
                    "class_users_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "joined_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "class_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    }
                },
                "type": "object"
            },
...
            "UserResponse": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "profile_picture": {
                        "properties": {
                            "file_id": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "url": {
                                "type": "string",
                                "format": "uri"
                            },
                            "filename": {
                                "type": "string"
                            },
                            "mime_type": {
                                "type": "string"
                            },
                            "size": {
                                "type": "integer"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}